post-list(feat):Change responsabilities
	- Move the 'Comments Feature' to PostContestList responsability
	- Remove 'NavigationStack' and create a view to separete other views

Please enter the commit message for your changes. Lines starting
