+++ catto/Catto/Features/Post/PostView.swift
@@ -11,40 +11,43 @@ struct PostView: View {
     let viewModel: PostViewModel
 
     var body: some View {
-        VStack(spacing: 16) {
-            CatImagesView(urlString: viewModel.post.cat.url)
-            if viewModel.post.caption.isEmpty {
-                CommentTextFieldView(viewModel: viewModel, text: "Title me if you can!")
-            } else {
-                ForEach(viewModel.post.caption, id: \.self) { eachCaption in
-                    MiauCard(
-                        username: eachCaption.user.name,
-                        avatarUrl: eachCaption.user.imageUrl,
-                        caption: eachCaption.caption,
-                        likes: eachCaption.vote,
-                        isMostVoted: false
-                    )
-                }
-                CommentTextFieldView(viewModel: viewModel, text: "Title me the best you can")
-            }
-        }
-        .padding([.top, .horizontal], 16)
-        .padding(.bottom, 24)
-        .background(.bgScreen)
-        .clipShape(RoundedRectangle(cornerRadius: 16))
-        .offset(x: viewModel.offset.width, y: viewModel.offset.height * 0.4)
-        .rotationEffect(.degrees(Double(viewModel.offset.width / 60)))
-        .gesture(
-            DragGesture()
-                .onChanged { value in
-                    viewModel.offset = value.translation
-                }
-                .onEnded { value in
-                    withAnimation(.spring()) {
-                        viewModel.swipeCard(width: viewModel.offset.width)
+        ScrollView(showsIndicators: false) {
+            VStack(spacing: 16) {
+                CatImagesView(urlString: viewModel.post.cat.url)
+                if viewModel.post.caption.isEmpty {
+                    Text("No comments yet!")
+                        .heading()
+                        .foregroundStyle(.bgCardSecondary)
+                } else {
+                    ForEach(viewModel.post.caption, id: \.self) { eachCaption in
+                        MiauCard(
+                            username: eachCaption.user.name,
+                            avatarUrl: eachCaption.user.imageUrl,
+                            caption: eachCaption.caption,
+                            likes: eachCaption.vote,
+                            isMostVoted: false
+                        )
                     }
                 }
-        )
+            }
+            .padding([.top, .horizontal], 16)
+            .padding(.bottom, 24)
+            .background(.bgScreen)
+            .clipShape(RoundedRectangle(cornerRadius: 16))
+            .offset(x: viewModel.offset.width, y: viewModel.offset.height * 0.4)
+            .rotationEffect(.degrees(Double(viewModel.offset.width / 60)))
+            .gesture(
+                DragGesture()
+                    .onChanged { value in
+                        viewModel.offset = value.translation
+                    }
+                    .onEnded { value in
+                        withAnimation(.spring()) {
+                            viewModel.swipeCard(width: viewModel.offset.width)
+                        }
+                    }
+            )
+        }
     }
 }
 
@@ -87,40 +90,5 @@ struct CatImageProgressView: View {
                 RoundedRectangle(cornerRadius: 16)
                     .stroke(.auxiliarCardCatBoard, lineWidth: 8)
             })
-    }
-}
-
-struct CommentTextFieldView: View {
-    @State var viewModel: PostViewModel
-    @State var text: String
-    @FocusState private var commentFieldIsFocused: Bool
-
-    var body: some View {
-        HStack {
-            TextField(
-                "",
-                text: $viewModel.captionInput,
-                prompt: Text(text)
-                    .foregroundStyle(.textPrimaryDark)
-            )
-            .miauTextField()
-            .focused($commentFieldIsFocused)
-            .textInputAutocapitalization(.sentences)
-            .keyboardType(.default)
-            .submitLabel(.send)
-            .onSubmit {
-                if !viewModel.captionInput.isEmpty {
-                    viewModel.addCommentButtonPressed()
-                    commentFieldIsFocused = false
-                }
-
-            }
-            MiauButtonSend {
-                if !viewModel.captionInput.isEmpty {
-                    viewModel.addCommentButtonPressed()
-                    commentFieldIsFocused = false
-                }
-            }
-        }
     }
 }
